Added basic add user to database provider
